Run hook after each test to truncate all DB tables.
This gives each test a "clean" database snapshot to work with.
Unfortunately this means we can't run tests in parallel or weird things
will happen. Seems like the easiest thing for now given knex and it's
abstraction for inserts and auto transaction wrapping.

In the future we could consider wrapping every single test in a
transaction which would allow for parallel tests in certain contexts.