remove tracked_token_account_balances references (quick and dirty)
this code is mostly unused and will likely be abandoned indefinitely, so
wasn't worth the effort to do a full rewrite. mainly commenting out references
to the table that weren't exposed anywhere publicly anyway (e.g. it was on the
API but not read by the UI anywhere)

the only significant logic changes were

1. using tracked_token_accounts instead of tracked_token_account_balances for
the totalWallets metric. we had to remove the option to filter by balance > 0
for this, but that functionality wasn't used anywhere

2. the getAllTrackedTokenAccountInfoAndTransactionsForEndDate cron now uses
the last fetched transaction instead of the last available balance. this should
functionality be the same